requirements.txt
Keras==2.4.3
pandas==1.0.5
keras_retinanet==1.0.0
opencv_python==4.3.0.36
numpy==1.18.2
tensorflow==2.2.0
Pillow==8.0.1
scikit_learn==0.23.2
